requirement.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# AI and Language Models
langchain==0.0.335
langchain-groq==0.0.1
openai==1.3.0
groq==0.4.0

# Database
pymongo==4.5.0
chromadb==0.4.18

# Data Processing
pandas==2.1.1
numpy==1.26.0

# Date and Time Handling
python-dateutil==2.8.2

# Caching and Performance
joblib==1.3.2

# Logging and Monitoring
loguru==0.7.2

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1